+++ src/main/java/de/samply/reporter/script/ScriptEngineManager.java
@@ -1,5 +1,6 @@
 package de.samply.reporter.script;
 
+import de.samply.reporter.context.CellContext;
 import de.samply.reporter.context.Context;
 import de.samply.reporter.template.ReportTemplate;
 import de.samply.reporter.template.script.Script;
@@ -7,6 +8,7 @@ import de.samply.reporter.template.script.ScriptFramework;
 import java.util.Arrays;
 import java.util.HashMap;
 import java.util.Map;
+import org.apache.poi.ss.usermodel.Workbook;
 import org.springframework.stereotype.Component;
 
 @Component
@@ -43,13 +45,21 @@ public class ScriptEngineManager {
   private ScriptResult generateRawResultWithoutExceptionHandling(Script script,
       Context context)
       throws ScriptEngineException {
+    return idScriptEngineMap.get(fetchScriptFramework(script)).generateRawResult(script, context);
+  }
+
+  private ScriptFramework fetchScriptFramework(Script script) {
     ScriptFramework scriptFramework =
         (script.getFramework() != null) ? ScriptFramework.valueOfFramework(script.getFramework())
             : ScriptFramework.getDefault();
     if (scriptFramework == null) {
       throw new RuntimeException("Script Framework " + script.getFramework() + " not found");
     }
-    return idScriptEngineMap.get(scriptFramework).generateRawResult(script, context);
+    return scriptFramework;
+  }
+
+  public CellContext generateCellContext(Script script, Workbook workbook) throws ScriptEngineException {
+    return idScriptEngineMap.get(fetchScriptFramework(script)).generateCellContext(script, workbook);
   }
 
 }